fix(fimframe): nages as integer
Thanks to @Andrea-Havron-NOAA for the suggestion while in
pull request #327

+++ R/fimsframe.R
@@ -24,7 +24,7 @@ setClass(
   Class = "FIMSFrameAge",
   slots = c(
     ages = "numeric",
-    nages = "numeric",
+    nages = "integer",
     weightatage = "data.frame"
   ),
   contains = "FIMSFrame"
